Bin Allocation Algorithms

WeightedEnsemble.trivial_allocation! — Function

trivial_allocation!: Trivially allocate each particle to have one offspring.

Arguments

E - particle ensemble

B - bin data structure

WeightedEnsemble.minimal_bin_allocation! — Function

minimal_bin_allocation!: Allocates a single particle to be spawned within each nonempty bin

Arguments

B - bin data structure

WeightedEnsemble.targeted_bin_allocation! — Function

targeted_bin_allocation!: Targeted allocation of particles amongst bins using a specified function, G:(p, E, B, t) → [0,∞) for bin p. Falls back to uniform allocation amongst the non-empty bins in the event that this G fails to normalize.

Arguments

B - bin data structure

E - particle ensemble

G - target function

t - t-th seletion step

n_allocate - number of particles to allocate
Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

WeightedEnsemble.optimal_bin_allocation! — Function

optimal_bin_allocation!: Optimally particles according to the bins, using a value function to approximate mutation variance.

Arguments

B - bin data structure

E - particle ensemble

v² - v² variance function estimator

t - t-th seletion step

n_allocate - number of particles to allocate

Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

WeightedEnsemble.uniform_bin_allocation! — Function

uniform_bin_allocation!: Uniformly allocate particles amongst bins.

Arguments

B - bin data structure

E - particle ensemble

n_allocate - number of particles to allocate

Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

WeightedEnsemble.within_bin_allocation! — Function

within_bin_allocation!: Once the number of offspring within each bin are set, allocate them amongst the particles within the bin. This assumes that the bin allocations of the bins have completed.

Arguments

E - particle ensemble

B - bin data structure

Optional Arguments

within_bin_resampler=multinomial - resampling scheme within bins