Coarse Models

Coarse Models

Serial Methods

Multithreaded Methods

Distributed Parellel Methods

Several tools are included for building the variance and discrepancy functions, discussed in [1–3, 5]. These methods require, first, the construction of a coarse scale transition matrix $\tilde{k}$. This typically corresponds to an approximation of the Markov kernel, $K$, on the user defined bins,

\[\tilde{K}_{pq} \approx \mathbb{P}(p\to q),\]

Having, computed this matrix, we next solve for the coarse scale discrepancy and one step mutation variance functions:

\[\begin{gather*} (I - \tilde{K})\tilde{h} = \tilde{f} - \tilde{\mu}(\tilde{f})\\ \tilde{v}^2(p) = \mathrm{Var}_{\tilde{K}_{p,\bullet}}(\tilde{h}) \end{gather*}\]

Serial Methods

WeightedEnsemble.build_coarse_transition_matrix — Function

build_coarse_transition_matrix(mutation!, bin_id, x0_vals, n_bins, n_trials)

Contruct a transition matrix amongst the bins (serial version).

Arguments

mutation! - an in place mutation function

bin_id - bin identification function

x0_vals - an array of starting values

bin0_vals - an array of the bins corresponding to x0_vals

n_bins - total number of bins

n_trials - number of independent trials for each x0 starting value

WeightedEnsemble.build_coarse_poisson — Function

build_coarse_poisson(K̃, f̃)

Construct the solution to the Poisson problem and the 1-step variance approximation on the coarser model given the transition matrix, K̃, and a coarse scale QoI function, f̃. This solves it using Julia's linear solver.

Arguments

K̃ - coarse scale transition matrix

f̃ - quantity of interest vector on the bin space

WeightedEnsemble.build_coarse_vectors — Function

build_coarse_vectors(n_we_steps, K̃, f̃)

Assemble the conditional expectation and 1- step variance approximations on a coarse model, given the transition matrix, K̃, and a coarse scale QoI function, f̃.

Arguments

n_we_steps - number of WE steps

K̃ - coarse scale transition matrix

f̃ - quantity of interest vector on the bin space

While build_coarse_poisson is appropriate when using WE with steady state problems, build_coarse_vectors is what should be invoked for finite time horizon problems; see [5].

Multithreaded Methods

TBW

Distributed Parellel Methods

TBW