Coarse Models

Serial Algorithms

WeightedEnsemble.build_coarse_transition_matrix — Function

build_coarse_transition_matrix: Contruct a transition matrix amongst the bins (serial version).

Arguments

mutation! - an in place mutation function

bin_id - bin identification function

x0_vals - an array of starting values

bin0_vals - an array of the bins corresponding to x0_vals

n_bins - total number of bins

n_trials - number of independent trials for each x0 starting value

WeightedEnsemble.build_coarse_vectors — Function

build_coarse_vectors: Assemble the conditional expectation and 1- step variance approximations on a coarser model, given the transition matrix, K̃, and a coarse scale QoI function, f̃.

Arguments

n_we_steps - number of WE steps

K̃ - coarse scale transition matrix

f̃ - quantity of interest vector on the bin space

WeightedEnsemble.build_coarse_poisson — Function

build_coarse_poisson: Construct the solution to the Poisson problem and the 1-step variance approximation on the coarser model given the transition matrix, K̃, and a coarse scale QoI function, f̃. This solves it using Julia's linear solver.

Arguments

K̃ - coarse scale transition matrix

f̃ - quantity of interest vector on the bin space

Parallel Algorithms

WeightedEnsemble.pbuild_coarse_transition_matrix — Function

pbuild_coarse_transition_matrix: Contruct a transition matrix amongst the bins in parallel. It assumed that a pool of workers has already been contructed. Returns a sparse matrix.

Arguments

mutation! - an in place mutation function

bin_id - bin identification function

x0_vals - an array of starting values

bin0_vals - an array of the bins corresponding to x0_vals

n_bins - total number of bins

n_trials - number of independent trials for each x0 starting value

Multithreaded Algorithms

WeightedEnsemble.tbuild_coarse_transition_matrix — Function

tbuild_coarse_transition_matrix: Contruct a transition matrix amongst the bins (multithreaded version).

Arguments

mutation! - an in place mutation function

bin_id - bin identification function

x0_vals - an array of starting values

bin0_vals - an array of the bins corresponding to x0_vals

n_bins - total number of bins

n_trials - number of independent trials for each x0 starting value