Particle Selection Algorithms

WeightedEnsemble.trivial_selection! — Function

trivial_selection!: Trivial selection, copying over particles

Arguments

E - particle ensemble

WeightedEnsemble.repopulate! — Function

repopulate!: After allocating the number of offspring of each particle, copy the particles over.

Arguments

E - particle ensemble

B - bin data structure

Optional Arguments

WeightedEnsemble.uniform_selection! — Function

uniform_selection!: Uniformly select particles, ensuring each bin with positive bin weight has at least one offspring.

Arguments

E - particle ensemble

B - bin data structure

t - t-th seletion step

Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

within_bin_resampler=multinomial - resampling scheme within bins

WeightedEnsemble.optimal_selection! — Function

optimal_selection!: Perform optimal selection of the particles, ensuring each non empty bin has at least one particle.

Arguments

E - particle ensemble

B - bin data structure

v² - v² variance function estimator

t - t-th seletion step

Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

within_bin_resampler=multinomial - resampling scheme within bins

WeightedEnsemble.targeted_selection! — Function

targeted_selection!: Perform targeted selection of the particles, ensuring each non empty bin has at least one particle.

Arguments

E - particle ensemble

B - bin data structure

G - target function

t - t-th seletion step

Optional Arguments

allocation_resampler=systematic - resampling scheme amongst bins

within_bin_resampler=multinomial - resampling scheme within bins