Utility Functions

Some additional functions have been provided that, while they are not directly used, may be of use to the user.

WeightedEnsemble.update_bin_weights! — Function

update_bin_weights!(B, E)

Update the bin weights and bin particle counts from the ensemble

Arguments

B - bin data structure

E - particle ensemble

WeightedEnsemble.trivial_analysis! — Function

trivial_analysis!(E, B, t)

Default, trivial, analysis! step.

WeightedEnsemble.Voronoi_bin_id — Function

Voronoi_bin_id(X, tree)

Convenience function for bin id in Voronoi based bins

Arguments

X - Point thats bin is to be determined

tree - A nearest neighbors tree structure constructed with KDTree

WeightedEnsemble.repopulate! — Function

repopulate!(E, B)

After allocating the number of offspring of each particle, copy the particles over. If the number of allocated particles is < N, the remaining particles are assigned zero weight.

Arguments

E - particle ensemble

B - bin data structure