Utility Functions

WeightedEnsemble.update_bin_weights! — Function

update_bin_weights!: Update the bin weights from the ensemble

Arguments

B - bin data structure

E - particle ensemble

WeightedEnsemble.Points_to_Bins — Function

Points_to_Bins: Convenience function for constructing a bin structure using a sequence of points

Arguments

points - An array of points defining the Voronoi cells

WeightedEnsemble.Voronoi_to_Bins — Function

Voronoi_to_Bins: Convenience function for constructing a bin structure using a sequence of points as the Voronoi sites

Arguments

sites - An array of points defining the Voronoi cells

WeightedEnsemble.Voronoi_bin_id — Function

Voronoi_bin_id: Convenience function for bin id in Voronoi based bins

Arguments

X - Point thats bin is to be determined tree - A nearest neighbors tree structure constructed with KDTree

WeightedEnsemble.setup_Voronoi_bins — Function

setup_Voronoi_bins - Convenience function for constructing bins, a bin id function, and a rebinning function based on a set of Voronoi points

Arguments

voronoi_pts - User specified Voronoi cell centers

WeightedEnsemble.Dirac_to_Ensemble — Function

Dirac_to_Ensemble: Convenience function for construction an ensemble from a single initial walker. This hard codes the weights to be Float64.

Arguments

X - Starting state of all walkers n_particles - Number of walkers in the ensemble

Missing docstring.

Missing docstring for WeightedEnsenble.trivial_analysis!. Check Documenter's build log for details.