Weighted Ensemble Methods

Weighted Ensemble Methods

Serial Methods

Multithreated Methods

Distributed Methods

Serial Methods

WeightedEnsemble.run_we — Function

run_we(E₀, B₀, sampler, n_we_steps; n_save_iters = 1)

Run a serial WE simulation, optionally returning the ensemble at each, step with

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

Optional Arguments

n_save_iters = 1 - save the ensemble and bins every n_save_iters iterations. Set n_save_iters=n_we_steps to only save the final values.

WeightedEnsemble.run_we_observables — Function

run_we_observables(E₀, B₀, sampler, n_we_steps, observables)

Run a serial WE simulation, returning the values a specified fucntion, f, along the trajecotry.

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

observables - Tuple of scalar observable functions for the ergodic average

WeightedEnsemble.run_we! — Function

run_we!(E, B, sampler, n_we_steps)

Run an in place serial WE simulation with

Arguments

E - particle ensemble

B - bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

Multithreated Methods

These methods make use of multithreading in the mutation step.

WeightedEnsemble.trun_we — Function

trun_we(E₀, B₀, sampler, n_we_steps; n_save_iters = 1)

Run a multithreaded WE simulation, optionally returning the ensemble at each, step with

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

Optional Arguments

n_save_iters = 1 - save the ensemble and bins every n_save_iters iterations. Set n_save_iters=n_we_steps to only save the final values.

WeightedEnsemble.trun_we_observables — Function

trun_we_observables(E₀, B₀, sampler, n_we_steps, observables)

Run a multithreaded WE simulation, returning the values a specified fucntions, observables, along the trajecotry.

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

observables - Tuple of scalar observable functions for the ergodic average

WeightedEnsemble.trun_we! — Function

trun_we!(E, B, sampler, n_we_steps)

Run an in place multithreaded WE simulation with

Arguments

E - particle ensemble

B - bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

Distributed Methods

These methods make use of distributed computation in the mutation step via pmap.

WeightedEnsemble.prun_we — Function

prun_we: Run a parallel WE simulation, optionally returning the ensemble at each step. This performs the mutation steps in parallel, and assumes a worker pool has already been created.

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

Optional Arguments

n_save_iters = 1 - save the ensemble and bins every n_save_iters iterations. Set n_save_iters=n_we_steps to only save the final values.

WeightedEnsemble.prun_we_observables — Function

prun_we_observables: Run a parallel WE simulation, optionally returning the ensemble at each step. This performs the mutation steps in parallel, and assumes a worker pool has already been created.

Arguments

E₀ - initial particle ensemble

B₀ - initial bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run

observables - Tuple of scalar observable functions for the ergodic average

WeightedEnsemble.prun_we! — Function

prun_we!: Run an in place parallel WE simulation. This performs the mutation steps in parallel, and assumes a worker pool has already been created.

Arguments

E - particle ensemble

B - bin data structure

sampler - WE sampler functions data structure

n_we_steps - number of steps in the WE run